requirements.txt

PyPDF2==3.0.1
PyQt5==5.15.6
PyQt5-sip==12.9.1
PyQtWebEngine==5.15.5
pyRFC3339==1.1
python-apt==2.4.0
python-dateutil==2.8.1
python-debian===0.1.43ubuntu1
python-telegram-bot==13.14
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
pyxdg==0.27
pyxnat==1.5
PyYAML==5.4.1
pyzmq==22.3.0
rdflib==6.2.0
reportlab==3.6.8
requests==2.25.1
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.9.3
scour==0.38.2
SecretStorage==3.3.1
sentry-sdk==1.4.3
simplejson==3.18.0
six==1.16.0
soupsieve==2.3.1
systemd-python==234
termcolor==2.2.0
tornado==6.1
traits==6.3.2
tzdata==2022.6
tzlocal==4.2
ubuntu-advantage-tools==27.12
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
uritemplate==3.0.1
urllib3==1.26.5
usb-creator==0.3.7
vboxapi==1.0
wadllib==1.3.6
webencodings==0.5.1
xdg==5
xkit==0.0.0
xlrd==2.0.1
xlwt==1.3.0
zipp==1.0.0
